Clean up language selector (#2460)
* Remove unused dependency

This was added during #2444 for seemingly no reason.

* Simplify ChapterSelect state update logic

* Fix sublanguage rendering logic and update types

Turns out the chapter list rendered has been used wrongly this whole
time, resulting in an inability to filter. After fixing, a lot of the
conditional rendering can be simplified. This is due to an additional
attribute of SALanguage that was created: the main language that a
sublanguage belongs to.

* Fix Scheme, Python not loading on SA @ X deploys

* Clean up code

* Remove commented code
* Remove console.log statement
* Remove useless React fragment

* Refactor language names to use enum

* Update test and snapshot

* Use the language enum in the component test
* Remove console.log from test file
* Update snapshot

* Migrate playground language state from to use enum

* Remove commented code

* Refactor sublanguage definitions

* Update broken tests

* Fix lint error

* Improve language abstractions

Removed unneeded exports. Also updated the default language selector to
match the abstraction level of the chapter selector in the Playground.