Create language configuration for conditional rendering of tabs (#2464)
* Create and persist Playground language config action

* Migrate information to language config

* Throw error if language config can't be found

* Hide incorrectly exported sourceSubLanguages

* Fix broken test

* Refactor language config keys

* Update language config on chapter select

* Use language config for env/subst visualizer rendering

* Abstract language features type and refactor logic

* Type the sublanguages

* Update comments

* Add a TODO
* Remove unused comments

* Add test for Source language features

* Migrate data visualizer rendering logic to config

* Refactor language config selector

* Update dependency array

* Migrate multi-file support to language config

Also creates a test snapshot for Source language configs.

* Add test snapshots for other languages

* Support non-Source languages in links

* Fix language configuration not loaded on refresh

Previously, setting the language to Source 1 and refreshing the page
will cause the folder mode toggle to reappear until Source 1 is
reselected. This fixes that issue.

* Setup languageConfig fallbacks

* Revert "Setup languageConfig fallbacks"

This reverts commit ceb543d46f31353e198c9e204f6a1b4642e1db50.

* Fix bugs and refactor

* Add guard clause

* Enforce presence of `supports` key for lang features

* Remove manual if-else in favor of language config

Removes the no-longer-needed if-else clauses to disable certain tabs
when using non-Source languages. Also migrates the logic of only
enabling the data visualiser tab when using full JS/TS into their
respective language configurations, rendering the corresponding if-else
clause unnecessary as well.

* Fix failing tests

* Remove unnecessary attribute access

Because of the enforced typing of the sublanguages, it is guaranteed
that there will not be any existing object in that attribute, so there
is no need to worry about accidentally overriding the supported
features.