Fix bugs involving stepper in stories (#2620)
* Use most recent output for stepper

* Remove early return

This caused issues with tab handling due to interactions with show/hide
behavior.

* Create custom styles for stories side content

* Wrap StoriesSideContent in card and custom div

Improves layout

* Remove tab hiding logic from StoriesSideContent

Done in favour of more intuitive new UI/UX.

* Move stories output to a side content tab

Done in order to allow for change tabs callback to run when toggling
between normal output mode and stepper output mode.

* Set default tab to stories output

* Fire tab change callback when run button is clicked

* Fix incorrect tab change behavior

After removing the early return, as well as due to the possibility of
having multiple instances of the component sharing the same environment,
some of the assumptions that were previously had are no longer valid.
Thus, the tab-update callback logic needs to be updated accordingly.

* Add TODO

* Use component state over env state for usingSubst

Previously, whether or not the `Output` component needs to interpret the
incoming output as something for the stepper was dictated by the
environment state. Since the currently showing tab depends on the
component, as opposed to the environment setting, this resulted in a
runtime error when the currently selected tab is not the stepper but the
output tries to interpret the stepper output as a normal output.

* Create separate storiesRun side content type

Resolves FIXME.

* Fix workspace tab indicator spacing

This regression was introduced when Blueprint was upgraded a while back.